+++ docs/conf.py
@@ -9,7 +9,7 @@
 project = "eBird Checklists"
 copyright = "2024, Stuart MacKay"
 author = "Stuart MacKay"
-release = "1.0.3"
+release = "1.1.0"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration